Correctly write last modified value on replace; better typing for leases on replace in Thrall
