+++ leases/app/lib/LeaseNotifier.scala
@@ -2,34 +2,9 @@ package lib
 
 import com.gu.mediaservice.lib.aws.{MessageSender, UpdateMessage}
 import com.gu.mediaservice.lib.formatting._
-import com.gu.mediaservice.model.{LeasesByMedia, MediaLease}
+import com.gu.mediaservice.model.{LeaseNotice, LeasesByMedia, MediaLease}
 import org.joda.time.DateTime
 import play.api.libs.json._
-
-case class LeaseNotice(mediaId: String, leaseByMedia: JsValue) {
-  def toJson = Json.obj(
-    "id" -> mediaId,
-    "data" -> leaseByMedia,
-    "lastModified" -> printDateTime(DateTime.now())
-  )
-}
-
-object LeaseNotice {
-  import JodaWrites._
-
-  implicit val writer = new Writes[LeasesByMedia] {
-    def writes(leaseByMedia: LeasesByMedia) = {
-      LeasesByMedia.toJson(
-        Json.toJson(leaseByMedia.leases),
-        Json.toJson(leaseByMedia.lastModified)
-      )
-    }
-  }
-  def apply(mediaLease: MediaLease): LeaseNotice = LeaseNotice(
-    mediaLease.mediaId,
-    Json.toJson(LeasesByMedia(List(mediaLease), Some(mediaLease.createdAt)))
-  )
-}
 
 class LeaseNotifier(config: LeasesConfig, store: LeaseStore) extends MessageSender(config, config.topicArn) {
   private def build(mediaId: String, leases: List[MediaLease] ): LeaseNotice = {
