+++ thrall/app/lib/ElasticSearch6.scala
@@ -285,9 +285,10 @@ class ElasticSearch6(config: ElasticSearch6Config, metrics: ThrallMetrics) exten
     List(eventualUpdateResponse.map(_ => ElasticSearchUpdateResponse()))
   }
 
-  def replaceImageLeases(id: String, leaseByMedia: JsLookupResult, lastModified: JsLookupResult)(implicit ex: ExecutionContext): List[Future[ElasticSearchUpdateResponse]] = {
+  def replaceImageLeases(id: String, leases: Seq[MediaLease])(implicit ex: ExecutionContext): List[Future[ElasticSearchUpdateResponse]] = {
     val replaceLeasesScript = loadPainless(
-    """ | ctx._source.leases = params.leaseByMedia;"""
+      """ | ctx._source.leases.lastModified = params.lastModified;
+        | ctx._source.leases.leases = params.leases;""".stripMargin
     )
 
     val scriptSource = loadPainless(s"""
@@ -295,19 +296,16 @@ class ElasticSearch6(config: ElasticSearch6Config, metrics: ThrallMetrics) exten
                                        |   $updateLastModifiedScript
                                        | """)
 
-    val lleaseByMediaParameter = leaseByMedia.toOption.map(asNestedMap)
-    val lastModifiedParameter = lastModified.toOption.map(_.as[String])
-
     val params = Map(
-      "leaseByMedia" -> lleaseByMediaParameter.getOrElse(null),
-      "lastModified" -> lastModifiedParameter.getOrElse(null)
+      "leases" -> Json.toJson(leases),
+      "lastModified" -> Json.toJson(DateTime.now.toString)
     )
 
     val script = Script(script = replaceLeasesScript).lang("painless").params(params)
 
     val updateRequest = updateById(imagesAlias, Mappings.dummyType, id).script(script)
 
-    val eventualUpdateResponse = executeAndLog(updateRequest, s"ES6 updating all leases on image $id with: $lleaseByMediaParameter")
+    val eventualUpdateResponse = executeAndLog(updateRequest, s"ES6 updating all leases on image $id with: ${leases.toString}")
       .incrementOnFailure(metrics.failedSyndicationRightsUpdates){case _ => true}
 
     List(eventualUpdateResponse.map(_ => ElasticSearchUpdateResponse()))
