+++ thrall/app/lib/ElasticSearchRouter.scala
@@ -1,5 +1,5 @@
 package lib
-import com.gu.mediaservice.model.{Image, Photoshoot, SyndicationRights}
+import com.gu.mediaservice.model.{Image, MediaLease, Photoshoot, SyndicationRights}
 import play.api.libs.json.{JsLookupResult, JsValue}
 
 import scala.concurrent.{ExecutionContext, Future}
@@ -23,8 +23,8 @@ class ElasticSearchRouter(versions: Seq[ElasticSearchVersion]) extends ElasticSe
 
   override def deleteSyndicationRights(id: String)(implicit ex: ExecutionContext): List[Future[ElasticSearchUpdateResponse]] = versions.map(_.deleteSyndicationRights(id)).head
 
-  override def replaceImageLeases(id: String, leaseByMedia: JsLookupResult, lastModified: JsLookupResult)(implicit ex: ExecutionContext): List[Future[ElasticSearchUpdateResponse]] =
-    versions.map(_.replaceImageLeases(id, leaseByMedia, lastModified)).head
+  override def replaceImageLeases(id: String, leases: Seq[MediaLease])(implicit ex: ExecutionContext): List[Future[ElasticSearchUpdateResponse]] =
+    versions.map(_.replaceImageLeases(id, leases)).head
 
   override def addImageLease(id: String, lease: JsLookupResult, lastModified: JsLookupResult)(implicit ex: ExecutionContext): List[Future[ElasticSearchUpdateResponse]] =
     versions.map(_.addImageLease(id, lease, lastModified)).head
